:sparkles: Add copynet model for NER/RE
