+++ seq2rel/common/util.py
@@ -2,6 +2,8 @@ ENT_TYPE = "<{}>"
 REL_START = "<{}>"
 REL_END = "</{}>"
 
+RANDOM_STATE = 42
+
 
 def sanitize(text: str, lowercase: bool = False) -> str:
     """Cleans text by removing whitespace, newlines and tabs and (optionally) lowercasing."""
